requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
langchain
langgraph
langchain-openai
langchain-aws==0.2.1
boto3==1.34.162
streamlit==1.38.0
